+++ RocpProject2/src/app/app.component.html
@@ -8,6 +8,7 @@
   <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
     <div class="navbar-nav">
             <a class="nav-item nav-link" routerLink="getbyId">Get Task</a>
+            <a class="nav-item nav-link" routerLink="create">Create Task</a>
             <a class="nav-item nav-link" routerLink="update">Update Task</a>
             <a class="nav-item nav-link" routerLink="delete">Delete</a>
