+++ RocpProject2/src/app/create/create.component.html
@@ -1,0 +1,8 @@
+
+<div>
+    <form [formGroup] = "todos">
+        <input type = "text" formControlName ="title">
+        <button type="button" class="btn btn-primary" (click)='postTodoEc2(todos)'>Create Todo</button>
+    </form>
+</div>
+
