+++ RocpProject2/src/app/create/create.component.ts
@@ -1,0 +1,32 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+import {ROCPService} from '../services/rocp.service';
+import {FormGroup, FormControl} from '@angular/forms';
+
+@Component({
+  selector: 'app-create',
+  templateUrl: './create.component.html',
+  styleUrls: ['./create.component.css']
+})
+export class CreateComponent implements OnInit {
+
+  constructor(private route: ActivatedRoute, private rocp: ROCPService) { }
+  todos = new FormGroup({
+    title: new FormControl('')
+    });
+  postTodoEc2(todoSub: FormGroup) {
+    let form = JSON.stringify(todoSub.value);
+    console.log(form);
+    console.log(todoSub);
+    this.rocp.postTodo(form).subscribe(
+      response => {
+        console.log('success');
+
+      }
+    );
+
+  }
+  ngOnInit(): void {
+  }
+
+}
