--- RocpProject2/src/app/todos/todos.component.html
@@ -1,1 +1,0 @@
-<p>todos works!</p>
